+++ src/Sulu/Component/Webspace/Loader/XmlFileLoader10.php
@@ -27,7 +27,6 @@ use Sulu\Component\Webspace\Loader\Exception\InvalidWebspaceDefaultLocalizationE
 use Sulu\Component\Webspace\Loader\Exception\InvalidWebspaceDefaultSegmentException;
 use Sulu\Component\Webspace\Loader\Exception\PortalDefaultLocalizationNotFoundException;
 use Sulu\Component\Webspace\Loader\Exception\WebspaceDefaultSegmentNotFoundException;
-use Sulu\Component\Webspace\Loader\Exception\WebspaceLocalizationNotUsedException;
 use Sulu\Component\Webspace\Navigation;
 use Sulu\Component\Webspace\NavigationContext;
 use Sulu\Component\Webspace\Portal;
@@ -164,7 +163,6 @@ class XmlFileLoader10 extends BaseXmlFileLoader
         $this->validateWebspaceDefaultLocalization();
         $this->validateDefaultPortalLocalization();
         $this->validateWebspaceDefaultSegment();
-        $this->validateLocalizations();
     }
 
     /**
@@ -648,43 +646,6 @@ class XmlFileLoader10 extends BaseXmlFileLoader
             } catch (InvalidDefaultLocalizationException $ex) {
                 throw new InvalidPortalDefaultLocalizationException($this->webspace, $portal);
             }
-        }
-    }
-
-    /**
-     * Validate that all localizations are used in the portals.
-     *
-     * @throws Exception\PortalDefaultLocalizationNotFoundException
-     * @throws Exception\InvalidPortalDefaultLocalizationException
-     */
-    protected function validateLocalizations()
-    {
-        $locales = array_unique(
-            array_map(
-                function (Localization $localization) {
-                    return $localization->getLocale(Localization::UNDERSCORE);
-                },
-                $this->webspace->getAllLocalizations()
-            )
-        );
-
-        $portalLocales = [];
-        foreach ($this->webspace->getPortals() as $portal) {
-            $portalLocales = array_merge(
-                $portalLocales,
-                array_map(
-                    function (Localization $localization) {
-                        return $localization->getLocale(Localization::UNDERSCORE);
-                    },
-                    $portal->getLocalizations()
-                )
-            );
-        }
-
-        $portalLocales = array_unique($portalLocales);
-
-        if (array_diff($locales, $portalLocales) || array_diff($portalLocales, $locales)) {
-            throw new WebspaceLocalizationNotUsedException($this->webspace);
         }
     }