+++ CHANGELOG.md
@@ -4,6 +4,7 @@ CHANGELOG for Sulu
 * dev-master
     * HOTFIX      #3265 [ContentBundle]         Fixed internal-link selection for pages
     * HOTFIX      #3273 [MarkupBundle]          Fixed multiline mismatch in HtmlExtractor
+    * ENHANCEMENT #3266 [ContentBundle]         Added locale parameter to teaser-selection-list
 
 * 1.5.1 (2017-03-16)
     * ENHANCEMENT #3256 [DocumentManagerBundle] Added VersionNotFoundException to fos_rest configuration